+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+###Changed
+- Improved Semantics for Gies Widgets [#1307](https://github.com/rokwire/illinois-app/issues/1307).
 
 ## [3.2.12] - 2022-02-03
 ### Added