+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-###Changed
+### Added
+- Canvas Notifications data model and UI [#1274](https://github.com/rokwire/illinois-app/issues/1274).
+### Changed
 - Improved Semantics for Gies Widgets [#1307](https://github.com/rokwire/illinois-app/issues/1307).
 
 ## [3.2.12] - 2022-02-03